# Speed up `RGBClassTransformer` by an order of magnitude