--- a/rastervision_core/rastervision/core/data/utils/misc.py
+++ b/rastervision_core/rastervision/core/data/utils/misc.py
@@ -66,7 +66,7 @@ def normalize_color(
                     f'but found {type(color)}.')
 
 
-def rgb_to_int_array(rgb_array):
+def rgb_to_int_array(rgb_array: np.ndarray) -> np.ndarray:
     r = np.array(rgb_array[:, :, 0], dtype=np.uint32) * (1 << 16)
     g = np.array(rgb_array[:, :, 1], dtype=np.uint32) * (1 << 8)
     b = np.array(rgb_array[:, :, 2], dtype=np.uint32) * (1 << 0)